+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### ÖZELLİKLER
 
+- server/postgresql: PostgreSQL 11 sürümüne yükselt
 - runtime/javascript: Nodejs 11.x sürümüne yükselt
 - runtime/common: wkhtmltox (eski adıyla wkhtmltopdf) kitaplığını ekle